+++ projects/js-packages/charts/src/components/bar-chart/bar-chart.tsx
@@ -126,24 +126,14 @@ const BarChartInternal: FC< BarChartProps > = ( {
 		totalPoints,
 	} );
 
-	const { resolveGroupColor } = useGlobalChartsContext();
-
-	const getColor = useCallback(
-		( seriesData: SeriesData, index: number ) =>
-			resolveGroupColor( {
-				group: seriesData.group,
-				index,
-				overrideColor: seriesData.options?.stroke,
-			} ),
-		[ resolveGroupColor ]
-	);
+	const { getElementStyles } = useGlobalChartsContext();
 
 	const getBarBackground = useCallback(
 		( index: number ) => () =>
 			withPatterns
 				? `url(#${ getPatternId( chartId, index ) })`
-				: getColor( dataSorted[ index ], index ),
-		[ withPatterns, getColor, dataSorted, chartId ]
+				: getElementStyles( { data: dataSorted[ index ], index } ).color,
+		[ withPatterns, getElementStyles, dataSorted, chartId ]
 	);
 
 	const renderDefaultTooltip = useCallback(
@@ -341,12 +331,15 @@ const BarChartInternal: FC< BarChartProps > = ( {
 						<>
 							<defs data-testid="bar-chart-patterns">
 								{ dataSorted.map( ( seriesData, index ) =>
-									renderPattern( index, getColor( seriesData, index ) )
+									renderPattern( index, getElementStyles( { data: seriesData, index } ).color )
 								) }
 							</defs>
 							<style>
 								{ dataSorted.map( ( seriesData, index ) =>
-									createPatternBorderStyle( index, getColor( seriesData, index ) )
+									createPatternBorderStyle(
+										index,
+										getElementStyles( { data: seriesData, index } ).color
+									)
 								) }
 							</style>
 						</>