+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [1.3.4] - 2025-09-22
+### Changed
+- Update dependencies.
+
+## [1.3.3] - 2025-09-19
+### Changed
+- Update package dependencies. [#45173] [#45229]
+
 ## [1.3.2] - 2025-09-08
 ### Changed
 - Update package dependencies. [#45027] [#45097]
@@ -1528,6 +1536,8 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[1.3.4]: https://github.com/Automattic/jetpack-components/compare/1.3.3...1.3.4
+[1.3.3]: https://github.com/Automattic/jetpack-components/compare/1.3.2...1.3.3
 [1.3.2]: https://github.com/Automattic/jetpack-components/compare/1.3.1...1.3.2
 [1.3.1]: https://github.com/Automattic/jetpack-components/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/Automattic/jetpack-components/compare/1.2.2...1.3.0