+++ projects/js-packages/connection/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### This is a list detailing changes for the Jetpack RNA Connection Component releases.
 
+## [1.4.11] - 2025-09-22
+### Changed
+- Update dependencies.
+
+## [1.4.10] - 2025-09-19
+### Changed
+- Update package dependencies. [#45229]
+
 ## [1.4.9] - 2025-09-15
 ### Changed
 - Internal updates.
@@ -1166,6 +1174,8 @@
 - `Main` and `ConnectUser` components added.
 - `JetpackRestApiClient` API client added.
 
+[1.4.11]: https://github.com/Automattic/jetpack-connection-js/compare/v1.4.10...v1.4.11
+[1.4.10]: https://github.com/Automattic/jetpack-connection-js/compare/v1.4.9...v1.4.10
 [1.4.9]: https://github.com/Automattic/jetpack-connection-js/compare/v1.4.8...v1.4.9
 [1.4.8]: https://github.com/Automattic/jetpack-connection-js/compare/v1.4.7...v1.4.8
 [1.4.7]: https://github.com/Automattic/jetpack-connection-js/compare/v1.4.6...v1.4.7