+++ projects/js-packages/i18n-loader-webpack-plugin/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.85] - 2025-09-19
+### Changed
+- Update package dependencies. [#45173]
+- Update package dependencies. [#45229]
+
 ## [2.0.84] - 2025-09-08
 ### Changed
 - Update package dependencies. [#45027]
@@ -365,6 +370,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
+[2.0.85]: https://github.com/Automattic/i18n-loader-webpack-plugin/compare/v2.0.84...v2.0.85
 [2.0.84]: https://github.com/Automattic/i18n-loader-webpack-plugin/compare/v2.0.83...v2.0.84
 [2.0.83]: https://github.com/Automattic/i18n-loader-webpack-plugin/compare/v2.0.82...v2.0.83
 [2.0.82]: https://github.com/Automattic/i18n-loader-webpack-plugin/compare/v2.0.81...v2.0.82