+++ projects/js-packages/idc/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ### This is a list detailing changes for the Jetpack RNA IDC package releases.
 
+## 1.0.27 - 2025-09-22
+### Changed
+- Update dependencies. [#44300]
+
+## 1.0.26 - 2025-09-19
+### Changed
+- Update package dependencies. [#45229]
+
 ## 1.0.25 - 2025-09-08
 ### Changed
 - Update package dependencies. [#45027]