+++ projects/js-packages/number-formatters/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.12] - 2025-09-19
+### Changed
+- Update package dependencies. [#45239]
+
 ## [1.0.11] - 2025-09-08
 ### Changed
 - Update package dependencies. [#45097]
@@ -77,6 +81,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release
 - Basic number formatting functionality
 
+[1.0.12]: https://github.com/Automattic/number-formatters/compare/1.0.11...1.0.12
 [1.0.11]: https://github.com/Automattic/number-formatters/compare/1.0.10...1.0.11
 [1.0.10]: https://github.com/Automattic/number-formatters/compare/1.0.9...1.0.10
 [1.0.9]: https://github.com/Automattic/number-formatters/compare/1.0.8...1.0.9