+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2025-09-22
+### Fixed
+- Fix lints. [#45242]
+
+## [1.4.0] - 2025-09-19
+### Changed
+- Update @wordpress/dataviews package from v7 to v9. [#45213]
+- Update package dependencies. [#45229]
+
 ## [1.3.1] - 2025-09-15
 ### Changed
 - Internal updates.
@@ -1364,6 +1373,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[1.4.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.4.0...v1.4.1
+[1.4.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.2.6...v1.3.0
 [1.2.6]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.2.5...v1.2.6