+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.7.10 - 2025-09-22
+### Added
+Add `watchOptions` to shared config. [#45235]
+
+## 3.7.9 - 2025-09-19
+### Changed
+- Update package dependencies. [#45240]
+
+## 3.7.8 - 2025-09-19
+### Changed
+- Update package dependencies. [#45173]
+- Update package dependencies. [#45229]
+
 ## 3.7.7 - 2025-09-08
 ### Changed
 - Update package dependencies. [#45027] [#45097]