+++ projects/packages/connection/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.18.10] - 2025-09-22
+### Changed
+- Update dependencies. [#44300]
+
+## [6.18.9] - 2025-09-19
+### Changed
+- Update package dependencies. [#45173] [#45229]
+
 ## [6.18.8] - 2025-09-16
 ### Fixed
 - Prevent PHP warnings when handling invalid user data. [#45198]
@@ -1591,6 +1599,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Separate the connection library into its own package.
 
+[6.18.10]: https://github.com/Automattic/jetpack-connection/compare/v6.18.9...v6.18.10
+[6.18.9]: https://github.com/Automattic/jetpack-connection/compare/v6.18.8...v6.18.9
 [6.18.8]: https://github.com/Automattic/jetpack-connection/compare/v6.18.7...v6.18.8
 [6.18.7]: https://github.com/Automattic/jetpack-connection/compare/v6.18.6...v6.18.7
 [6.18.6]: https://github.com/Automattic/jetpack-connection/compare/v6.18.5...v6.18.6